+++ src/index.ts
@@ -1,1 +1,1 @@
-export * from './number-utils';
+export * from './lib';