+++ src/lib/clamp.test.ts
@@ -1,0 +1,93 @@
+import { describe, expect, it } from '@jest/globals';
+import { clamp } from './clamp';
+
+describe('clamp', () => {
+  describe('clamp()', () => {
+    interface Example {
+      readonly input: {
+        readonly value: number;
+        readonly min: number;
+        readonly max: number;
+      };
+      readonly expected: number;
+    }
+
+    const EXAMPLES: readonly Example[] = [
+      {
+        input: {
+          value: 0,
+          min: 0,
+          max: 0,
+        },
+        expected: 0,
+      },
+      {
+        input: {
+          value: 0,
+          min: 0,
+          max: 10,
+        },
+        expected: 0,
+      },
+      {
+        input: {
+          value: 1,
+          min: 0,
+          max: 10,
+        },
+        expected: 1,
+      },
+      {
+        input: {
+          value: 10,
+          min: 0,
+          max: 10,
+        },
+        expected: 10,
+      },
+      {
+        input: {
+          value: 0,
+          min: 1,
+          max: 10,
+        },
+        expected: 1,
+      },
+      {
+        input: {
+          value: 0,
+          min: -1,
+          max: 10,
+        },
+        expected: 0,
+      },
+      {
+        input: {
+          value: -2,
+          min: -1,
+          max: 10,
+        },
+        expected: -1,
+      },
+      {
+        input: {
+          value: 20,
+          min: 0,
+          max: 10,
+        },
+        expected: 10,
+      },
+    ];
+
+    for (const example of EXAMPLES) {
+      it(JSON.stringify(example), () => {
+        const actual = clamp(
+          example.input.value,
+          example.input.min,
+          example.input.max,
+        );
+        expect(actual).toEqual(example.expected);
+      });
+    }
+  });
+});