+++ src/lib/index.ts
@@ -1,0 +1,6 @@
+export * from './clamp';
+export * from './pad';
+export * from './parse-float';
+export * from './parse-integer';
+export * from './random-integer';
+export * from './round';